Change vanity listener to use message created event rather than typing started. (#20)
* Fix modmail post-save hook to reliably fetch DM channel

The post-save hook was fetching the initial message via user.dmChannel, which only works if the DM channel is already cached. 'dmChannel' can be null and the fetch fails, causing the initial modmail message to never be forwarded. This change uses 'createDM()' to ensure the DM channel is retrieved from Discord if not cached. This method is idempotent, so no duplicate channels are created, and initial messages are now sent correctly.

* Update dead link

Changes the dead rapple.xyz link to a Discord guild invite.

* Add VanityEvent

* missed roles

* separate from event module

* Change to message created event

the typing event is far too spotty - change to message create for a better guarantee of receiving the message.